requirements.txt
pandas
SQLAlchemy
psycopg2-binary
confluent_kafka
apache-airflow